# U2-M8-Challenge-Shevach-Joshua/tasker-service/src/main/java/com/trilogyed/tasker/model/CustomErrorResponse.java
package com.trilogyed.tasker.model;

import com.fasterxml.jackson.annotation.JsonFormat;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.NonNull;
import lombok.RequiredArgsConstructor;

import java.time.LocalDateTime;

@Data
@NoArgsConstructor
@RequiredArgsConstructor
public class CustomErrorResponse {

    @NonNull
    String errorCode;
    int status;
    @NonNull
    String errorMsg;
    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd hh:mm:ss")
    LocalDateTime timestamp;

}